Retire the `service` modules
