+++ src/prelude.rs
@@ -12,8 +12,7 @@ pub use crate::client;
 #[cfg(feature = "rtu")]
 pub mod rtu {
     pub use crate::client::rtu::*;
-    pub use crate::frame::rtu::RequestContext;
-    pub use crate::service::rtu::{Client, ClientContext};
+    pub use crate::rtu::{Client, ClientContext, RequestContext};
 }
 
 #[allow(missing_docs)]
